+++ package.json
@@ -13,7 +13,7 @@
     "test:unit": "./scripts/unit-test.sh"
   },
   "dependencies": {
-    "@prisma/client": "^4.3.1",
+    "@prisma/client": "^4.4.0",
     "@remix-run/express": "^1.6.8",
     "@remix-run/node": "^1.6.8",
     "@remix-run/react": "^1.6.8",
